+++ CHANGELOG.md
@@ -112,6 +112,7 @@
 - [9558](https://github.com/vegaprotocol/vega/issues/9558) - Feature tests for relative return metric transfers and reward
 - [9559](https://github.com/vegaprotocol/vega/issues/9559) - Feature tests for return volatility metric transfers and reward
 - [9564](https://github.com/vegaprotocol/vega/issues/9564) - Fix error message for too many staking tiers. 
+- [8421](https://github.com/vegaprotocol/vega/issues/8421) - Markets that spend too long in opening auction should be cancelled.
 
 ### 🐛 Fixes